+++ packages/keepkit-ui/src/KeepSyncFeedbackObserver.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import { useKeepSyncFeedback } from "./hooks/useKeepSyncFeedback";
+
+export function KeepSyncFeedbackObserver() {
+  useKeepSyncFeedback();
+  return null;
+}